base form of logging

+++ backend/leanstore/storage/buffer-manager/BufferManager.cpp
@@ -580,7 +580,7 @@ BufferFrame& BufferManager::resolveSwip(Guard& swip_guard, Swip<BufferFrame>& sw
     COUNTERS_BLOCK()
     {
       WorkerCounters::myCounters().dt_misses_counter[bf.page.dt_id]++;
-      if (FLAGS_trace_dt_id >= 0 && bf.page.dt_id == static_cast<u64>(FLAGS_trace_dt_id) &&
+      if (FLAGS_trace_dt_id >= 0 && bf.page.dt_id == FLAGS_trace_dt_id &&
           utils::RandomGenerator::getRand<u64>(0, FLAGS_trace_trigger_probability) == 0) {
         utils::printBackTrace();
       }
